0.13.105 fix WebElementExtended.scroll

+++ appium_extended_web_element/web_element_extended.py
@@ -405,7 +405,7 @@ class WebElementExtended(WebElementClick,
 
     def scroll_until_find(self,
                           locator: Union[Tuple, 'WebElementExtended', Dict[str, str], str],
-                          timeout_method: int = 120,
+                          timeout_method: float = 120.0,
                           contains: bool = True
                           ) -> Union['WebElementExtended', None]:
         """
